requirements.txt
aiofiles
aiohttp
asyncio
uvloop; sys_platform != 'win32'
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx
motor
pillow
psutil
py-tgcalls==2.2.8
pyrofork
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
unidecode
yt-dlp
youtube-search
youtube-search-python
flask_restful
flask
gunicorn
